refactor(transport): extract Zenoh publisher path into its own file
Moves createPublisher, the publishers-dictionary helpers, and the
ZenohTransportPublisher concrete class into a focused extension file.
Renames the internal NSLock property to publishersLock for clarity.
Public API is unchanged.
